Improve TEI recovery as described in Q.921 section 5.3 when all TEI's are exhausted.
tei-check2.patch
This patch will only send an identity check request for the group-TEI (127)
instead of 63 separate messages, which is more efficient.

How it works: all L2 entity TEI state machines change to ST_TEI_IDVERIFY
without sending a check request msg, so they can react to the id check
responses.  Then the id check request is started for the L2 entity with
TEI 127.

TE's may answer with more than one Ai in the id check response, so take
care of this too.

tei-check3.patch
It is not necessary to actually send an ID remove for the TEI's that
get no check response.  Sending this is left out when the check request
is done for TEI 127.

Patches:
    tei-check2.patch file uploaded by customer (with minor cosmetic changes).
    tei-check3.patch file uploaded by customer (with minor cosmetic changes).

JIRA ABE-1865