Fixed mISDNuser to handle Q.931 section 5.2.5.3.
This issue is related to Mantis 10457, which I think was never actually
fixed.  If you look closely you'll see that it's the same as described
here.

If there is more than one phone or one which occupies more than one TEI
(e.g.  Gigaset CX475) connected to an NT-PTMP port, the NT does not handle
the DISCONNECT messages from a phone correctly as required in Q.931
section 5.2.5.3.  Instead of releasing the call processes only for the TEI
that answered with DISCONNECT and retaining the cause until all TEIs have
finished their call processes, the L3 of mISDNuser passes it directly to
the call control (chan_misdn), which then hangs up the caller.  The call
processes to any other TEIs are not cleared and so the other phones all
keep ringing.

The case described in Mantis 10475 is with a Gigaset.  Gigasets use
separate TEIs for each handset, and if one handset answers the call, the
other TEIs are actively DISCONNECTed by the Gigaset, which then directly
leads to the hangup described in Mantis 10475.  The wireshark trace there
shows two TEIs, one connecting, the other one disconnecting the call.

Another symptom of this issue is if two (real) phones are connected to the
NT, the call (leg A) is hungup when just one of them rejects or is simply
busy.  The other one keeps ringing.

The this patch adds call clearing during call establishment according to
Q.931 section 5.2.5.3 to mISDNuser, which fixes both symptoms above.

JIRA ABE-1917
Patches:
    disconnect.patch (Modified to fix a potential memory leak and compiler warning.)