The handling of call transfer signaling for mISDN PTMP is not fully implemented.
The handling of call transfer signaling for mISDN PTMP is not fully
implemented.  The signaling of number updates with ISDN/DSS1 ECT
supplementary services (ETS 300 369-1) comes along with a notification
indicator IE and redirection number IE for PTMP.  The implementation in
the current Asterisk mISDN channel unfortunately can handle these
information elements only in a NOTIFY message.  These information elements
are also signaled in a FACILTY message with a RequestSubaddress facility,
when the subscriber is already in the active state (see 9.2.4 and 9.2.5 of
ETS 300 369-1).

**********

abe_2526_ast.patch

* Added support to handle the notification indicator IE and redirection
number IE with the RequestSubaddress facility.

* Made misdn_update_connected_line() send a NOTIFY message if Asterisk
originated the call and it is not connected yet.

* Made misdn_update_connected_line() send a FACILITY message if the call
is already connected.

This patch requires the presence of the associated mISDN patches to
compile.  I had to enhance mISDN to allow the notification indicator IE
and the redirection number IE to be used with a FACILITY message.  Earlier
versions of the Digium enhanced mISDN are no longer going to work.

**********

abe_2526_misdn.patch

* Made an incoming FACILITY message allow the presence of the notification
indicator IE and the redirection number IE.

**********

abe_2526_misdnuser_v3.patch

* Added support to send and receive a FACILITY message with the
notification indicator IE and the redirection number IE.

* Added the ability to send a NOTIFY message in PTMP/NT mode to all
responding subcalls in Q.931 states 6, 7, 8, 9, and 25.

**********

Patches:
      abe_2526_ast.patch uploaded by rmudgett (license 664)
      abe_2526_misdn.patch uploaded by rmudgett (license 664)
      abe_2526_misdnuser_v3.patch uploaded by rmudgett (license 664)
Tested by: rmudgett and reporter

JIRA SWP-2146
JIRA ABE-2526