Fix loss of available chan_misdn procids in NT PTMP mode.
There are two bugs in mISDNuser lib which eat up the available procids of
chan_misdn when it services a BRI NT PTMP port.  When this happens 256
times, it's not possible anymore to call out to a phone connected to the
NT port until asterisk is restarted.

The misdn driver stack normally sends a CC_RELEASE_CR to chan_misdn call
control when a call is completely released.  chan_misdn will then release
the used call process ID (procid).  But there are two situations where
this does not happen.

Case 1:
When a phone on NT is called and the called party rejects the call with a
cause, CC_RELEASE_CR is not generated.  In this case CC_RELEASE_COMPLETE
indication is also sent twice, which makes no sense at all.

Case 2:
When a phone on NT is called, and the calling party hangs up before the
call is answered, but after timer T312 (6s) has expired, CC_RELEASE_CR is
not generated due to a wrong 3rd argument to send_proc().

JIRA ABE-2811